+++ src/Facades/Bulksms.php
@@ -5,7 +5,9 @@ namespace Eugenefvdm\Api\Facades;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @method static mixed sendSMS(string $message, array $recipients)
+ * @method static \Eugenefvdm\Api\Bulksms encoding(string $encoding)
+ * @method static array sendSms(string $message, string|array $recipients)
+ * @method static array sendUnicodeSms(string $message, string|array $recipients)
  *
  * @see \Eugenefvdm\Api\Bulksms
  */